+++ server/src/services/password.service.ts
@@ -1,4 +1,5 @@
 import argon2 from "argon2";
+import crypto from "crypto";
 
 
 export class PasswordService {
@@ -9,4 +10,10 @@ export class PasswordService {
   async verifyPassword(hash: string, password: string): Promise<boolean> {
     return argon2.verify(hash, password);
   }
+
+  async generateOAuthPassword(): Promise<string> {
+    const randomPassword = crypto.randomBytes(32).toString('hex');
+    return await argon2.hash(randomPassword);
+  }
+
 }